+++ .vscode/settings.json
@@ -225,7 +225,9 @@
     "krekel",
     "libjpeg",
     "linalg",
+    "lljpeg",
     "logfile",
+    "losslessjpeg",
     "lzma",
     "macos",
     "magick",
