+++ pymedphys/_streamlit/apps/anonymise-monaco.py
@@ -24,12 +24,12 @@ import tempfile
 
 import streamlit as st
 
-from pymedphys._streamlit import misc as st_misc
-from pymedphys._streamlit import monaco as st_monaco
-from pymedphys._streamlit import rerun as st_rerun
+from pymedphys._streamlit.utilities import misc as st_misc
+from pymedphys._streamlit.utilities import monaco as st_monaco
+from pymedphys._streamlit.utilities import rerun as st_rerun
 
 HERE = pathlib.Path(__file__).parent.resolve()
-ANON_DEMOGRAPHIC_FILE = HERE.joinpath("demographic.000000")
+ANON_DEMOGRAPHIC_FILE = HERE.joinpath("data", "demographic.000000")
 
 
 "# Anonymise Monaco Files"
