+++ pymedphys/_streamlit/apps/mudensity/_dicom.py
@@ -17,8 +17,9 @@ from pymedphys._imports import streamlit as st
 
 import pymedphys
 from pymedphys._dicom.constants.uuid import DICOM_PLAN_UID
-from pymedphys._gui.streamlit.mudensity import _config, _exceptions
-from pymedphys._streamlit import misc as st_misc
+from pymedphys._streamlit.apps.mudensity import _config
+from pymedphys._streamlit.utilities import exceptions as _exceptions
+from pymedphys._streamlit.utilities import misc as st_misc
 from pymedphys._utilities import patient as utl_patient
