+++ pymedphys/_streamlit/apps/mudensity/_icom.py
@@ -19,12 +19,8 @@ import pathlib
 from pymedphys._imports import pandas as pd
 from pymedphys._imports import streamlit as st
 
-from pymedphys._gui.streamlit.mudensity import (
-    _config,
-    _deliveries,
-    _exceptions,
-    _utilities,
-)
+from pymedphys._streamlit.apps.mudensity import _config, _deliveries, _utilities
+from pymedphys._streamlit.utilities import exceptions as _exceptions
 from pymedphys._utilities import patient as utl_patient
