+++ pymedphys/_streamlit/utilities/exceptions.py
@@ -1,2 +1,26 @@
 class NoRecordsFound(ValueError):
     pass
+
+
+class InputRequired(ValueError):
+    pass
+
+
+class WrongFileType(ValueError):
+    pass
+
+
+class UnableToCreatePDF(ValueError):
+    pass
+
+
+class NoControlPointsFound(ValueError):
+    pass
+
+
+class NoMosaiqAccess(ValueError):
+    pass
+
+
+class ConfigMissing(ValueError):
+    pass
